New options for controlling error handling
Before this change, c2ffi command would return a successful exit code
even if compilation errors occurred. Add a new option --fail-on-error
which causes c2ffi to report a failure exit code if any compilation
error occurs. To maintain backwards compatibility, the current behaviour
is maintained unchanged when the option is not supplied.

This can be useful when c2ffi is invoked in build scripts, so the build
is aborted if an error unexpectedly occurs. While errors can be detected
by parsing stderr, it is far easier to be able to just look at the
return code.

Also add --warn-as-error option which tells clang to treat all warnings
as errors. This is the equivalent of passing -Werror on the clang
command line.