New --nostdinc option
With this option, standard header files such as /usr/include are not
included in the search path. This necessary if you want to use c2ffi in
a cross-compilation scenario, e.g. processing Windows headers while
running on Linux. Without this, it will mix the local platform headers
with the foreign platform headers, which is an invalid combination which
has a high likelihood of resulting in an error.