Add macOS 10.15 build to CI
Note this build workflow lies a bit - it tells CMake we are building
with GCC (`-DTOOLCHAIN=gcc`), but actually on macOS by default gcc is a
wrapper for AppleClang. I did it this way because it was the simplest
way to get it to work. Passing `-DTOOLCHAIN=clang` fails with a bunch of
errors due to clang not liking the parameters that CMake is passing.