Add --framework, --sys-framework options
These can be used to specify the user and system framework directories.
Frameworks are primarily a macOS concept (originally inherited from
NeXTSTEP), so these options are likely to be mainly useful on macOS (and
other closely related platforms such as iOS).

The most common use case is if you want to process headers for system
frameworks provided with macOS. In that case you want to pass something
like this:
```
--sys-framework $(xcrun --show-sdk-path)/System/Library/Frameworks
```

--framework is equivalent to the -F clang option.

--sys-framework is equivalent to the -iframework clang option.